added complete authentication using user_id and added route grouping also created auth.php and integrated with service provider

+++ app/Http/Controllers/IdeasController.php
@@ -12,30 +12,43 @@ class IdeasController extends Controller
         return view('ideas.show', compact('idea'));
     }
     
-    public function edit(Idea $idea){
-
-        $editing = true;
-        return view('ideas.show', compact('idea','editing'));
-
-    }
 
     public function store(){
         $validated = request()->validate([
             'content' => 'required|min:5|max:300',
         ]);
 
+        $validated['user_id'] = auth()->id();
         Idea::create($validated);
 
         return redirect()->route('homepage')->with('sucess','idea created Sucessfully!');
     }
+
+    public function edit(Idea $idea){
+
+        if (auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+        $editing = true;
+        return view('ideas.show', compact('idea','editing'));
+
+    }
+
     public function destroy(Idea $idea){
-        
+        if (auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()->route('homepage')->with('sucess','Idea deleted Sucessfully !');
     }
 
     public function update(Idea $idea){
+        if (auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+        
         $validated = request()->validate([
             'content' => 'required|min:10|max:300',
         ]);
